+++ oap-server/server-storage-plugin/storage-elasticsearch-plugin/src/test/java/org/apache/skywalking/oap/server/storage/plugin/elasticsearch/base/ElasticSearchColumnTypeMappingTestCase.java
@@ -32,26 +32,28 @@ public class ElasticSearchColumnTypeMappingTestCase {
     @Test
     public void test() throws NoSuchFieldException {
         ColumnTypeEsMapping mapping = new ColumnTypeEsMapping();
+        int defaultLength = 200;
 
-        Assertions.assertEquals("integer", mapping.transform(int.class, int.class, null));
-        Assertions.assertEquals("integer", mapping.transform(Integer.class, Integer.class, null));
+        Assertions.assertEquals("integer", mapping.transform(int.class, int.class, defaultLength, null));
+        Assertions.assertEquals("integer", mapping.transform(Integer.class, Integer.class, defaultLength, null));
 
-        Assertions.assertEquals("long", mapping.transform(long.class, long.class, null));
-        Assertions.assertEquals("long", mapping.transform(Long.class, Long.class, null));
+        Assertions.assertEquals("long", mapping.transform(long.class, long.class, defaultLength, null));
+        Assertions.assertEquals("long", mapping.transform(Long.class, Long.class, defaultLength, null));
 
-        Assertions.assertEquals("double", mapping.transform(double.class, double.class, null));
-        Assertions.assertEquals("double", mapping.transform(Double.class, Double.class, null));
+        Assertions.assertEquals("double", mapping.transform(double.class, double.class, defaultLength, null));
+        Assertions.assertEquals("double", mapping.transform(Double.class, Double.class, defaultLength, null));
 
-        Assertions.assertEquals("keyword", mapping.transform(String.class, String.class, null));
+        Assertions.assertEquals("keyword", mapping.transform(String.class, String.class, defaultLength, null));
+        Assertions.assertEquals("text", mapping.transform(String.class, String.class, 100_000, null));
 
         final Type listFieldType = this.getClass().getField("a").getGenericType();
-        Assertions.assertEquals("keyword", mapping.transform(List.class, listFieldType,
+        Assertions.assertEquals("keyword", mapping.transform(List.class, listFieldType, defaultLength,
                                                          new ElasticSearchExtension(null, null, false, false)
         ));
 
-        Assertions.assertEquals("keyword", mapping.transform(IntList.class, int.class,
+        Assertions.assertEquals("keyword", mapping.transform(IntList.class, int.class, defaultLength,
                                                          new ElasticSearchExtension(null, null, true, false)));
-        Assertions.assertEquals("text", mapping.transform(IntList.class, int.class,
+        Assertions.assertEquals("text", mapping.transform(IntList.class, int.class, defaultLength,
                                                          new ElasticSearchExtension(null, null, false, false)));
     }
 }